FrontEnd Validations
FrontEnd Validations:
Donate form
Request Form
Registration Form

Fixed profile 'choose file' button to take in images. The issue is that you won't be able to tell you uploaded an image (because the file name doesn't show up anymore....) we can remove the button styling, if this is important for UX.  Otherwise, the button is centered!